Eutech/Project/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
python-multipart
google-generativeai